+++ composenav/src/main/java/com/adamkobus/compose/navigation/di/NavModule.kt
@@ -1,9 +1,0 @@
-package com.adamkobus.compose.navigation.di
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@InstallIn(SingletonComponent::class)
-@Module(includes = [NavModuleInternalBinds::class])
-object NavModule