+++ composenav/src/main/java/com/adamkobus/compose/navigation/model/NavigationConsumerImpl.kt
@@ -1,14 +1,18 @@
 package com.adamkobus.compose.navigation.model
 
+import com.adamkobus.compose.navigation.ComposeNavigation
 import com.adamkobus.compose.navigation.NavigationConsumer
 import com.adamkobus.compose.navigation.action.NavAction
 import com.adamkobus.compose.navigation.action.NavActionWrapper
+import com.adamkobus.compose.navigation.data.NavigationResult
 import com.adamkobus.compose.navigation.intent.NavIntent
-import javax.inject.Inject
+import kotlinx.coroutines.CompletableDeferred
 
-internal class NavigationConsumerImpl @Inject constructor(
-    private val navProcessor: NavigationProcessor
-) : NavigationConsumer {
+internal class NavigationConsumerImpl : NavigationConsumer {
+
+    private val navProcessor
+        get() = ComposeNavigation.getNavigationProcessor()
+
     override fun offer(action: NavAction) {
         navProcessor.postNavAction(action)
     }
@@ -19,5 +23,23 @@ internal class NavigationConsumerImpl @Inject constructor(
 
     override fun offer(intent: NavIntent) {
         navProcessor.postNavIntent(intent)
+    }
+
+    override suspend fun offerBlocking(action: NavAction): NavigationResult {
+        val deferred = CompletableDeferred<NavigationResult>()
+        navProcessor.postNavAction(action, onTaskCompleted = deferred)
+        return deferred.await()
+    }
+
+    override suspend fun offerBlocking(action: NavActionWrapper): NavigationResult {
+        val deferred = CompletableDeferred<NavigationResult>()
+        navProcessor.postNavAction(action.action, onTaskCompleted = deferred)
+        return deferred.await()
+    }
+
+    override suspend fun offerBlocking(intent: NavIntent): NavigationResult {
+        val deferred = CompletableDeferred<NavigationResult>()
+        navProcessor.postNavIntent(intent, onTaskCompleted = deferred)
+        return deferred.await()
     }
 }