+++ composenav/src/main/java/com/adamkobus/compose/navigation/ui/NavComposable.kt
@@ -3,7 +3,7 @@ package com.adamkobus.compose.navigation.ui
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
-import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import com.adamkobus.compose.navigation.data.NavGraph
 
@@ -18,7 +18,7 @@ fun NavComposable(
     navController: NavHostController,
     vararg observedGraphs: NavGraph
 ) {
-    val vm = hiltViewModel<NavComposableVM>()
+    val vm: NavComposableVM = viewModel()
     LifecycleAwareComponent(vm)
     vm.graphsParam.bind(observedGraphs.toList())