+++ demo-common-ui/src/main/java/com/adamkobus/compose/navigation/demo/ui/nav/DogsBrowserGraph.kt
@@ -14,7 +14,7 @@ object DogsBrowserGraph : NavGraph("dogs") {
     val DogDetails = DogsList.next {
         param(PARAM_DOG_ID)
     }
-    val DemoDialog = dialogDestination("demoDialog")
+    val DemoDialog = navDestination("demoDialog")
     val Back = popDestination()
 }