+++ tutorial/src/main/java/com/adamkobus/compose/navigation/tutorial/nav/TutorialNavActions.kt
@@ -5,4 +5,9 @@ object TutorialNavActions {
     val FromWelcomeToList = TutorialGraph.Welcome goTo TutorialGraph.List
 
     fun fromListToDetail(itemId: Int) = TutorialGraph.List goTo TutorialGraph.Detail arg itemId
+
+    val FromDetailNavigateBack = TutorialGraph.Detail pop TutorialGraph.Back
+    fun fromDetailToDialog(itemId: Int) = TutorialGraph.Detail goTo TutorialGraph.DetailDialog arg itemId
+
+    val FromImageToDialog = TutorialGraph.Image goTo TutorialGraph.DetailDialog arg 1
 }